Add box around header line
Otherwise highlighted on the first line is indistinguishable from the header
line. The problem with adding the box is that on the right hand side it looks
as if the header line didn't extend all the way, but this is something I'm
going to have to live with for now.